完善配置界面功能，增加保存配置功能，树节点增加拖动，删除节点功能；调整采集卡配置界面布局

+++ src/Plugins/Astra.Plugins.DataAcquisition/ViewModels/DataAcquisitionDeviceDebugViewModel.cs
@@ -68,7 +68,7 @@ namespace Astra.Plugins.DataAcquisition.ViewModels
         public string DeviceName => _device?.DeviceName ?? "未知设备";
         public string DeviceId => _device?.DeviceId ?? string.Empty;
         public string SerialNumber => (_device?.CurrentConfig as DataAcquisitionConfig)?.SerialNumber ?? string.Empty;
-        public int SampleRate => (_device?.CurrentConfig as DataAcquisitionConfig)?.SampleRate ?? 0;
+        public double SampleRate => (_device?.CurrentConfig as DataAcquisitionConfig)?.SampleRate ?? 0.0;
         public int ChannelCount => (_device?.CurrentConfig as DataAcquisitionConfig)?.ChannelCount ?? 0;
         public int BufferSize => (_device?.CurrentConfig as DataAcquisitionConfig)?.BufferSize ?? 0;
